Moved doc fragment into a sub-docs fragment
